moved methods verifyAllProductsPrices() and verifyProductCost() to MainPageBL
