+++ src/main/java/com/opencart/steps/HeaderPageBL.java
@@ -65,52 +65,6 @@ public class HeaderPageBL {
         return this;
     }
 
-    /**
-     * Method finds all ProductContainers on the Page and checks
-     * if their Product Cost in specified currency is correct.
-     *
-     * @param currencyName is name of Currency that needs to be checked.
-     */
-
-    public void verifyAllProductsPrices(String currencyName) {
-        List<ProductContainer> productContainers = new SearchResultPage().getProductContainers();
-        for (ProductContainer container
-                : productContainers) {
-            verifyProductCost(container, currencyName);
-        }
-    }
-
-    /**
-     * Method checks if MacBook Cost in specified currency is correct.
-     *
-     * @param container    is Container for Products where price needs to be checked;
-     * @param currencyName is name of Currency that needs to be checked.
-     */
-    private void verifyProductCost(ProductContainer container, String currencyName) {
-        new ProductModel();
-        double expectedCost = ProductModel.getProductPrice(container.getName());
-        switch (currencyName.toLowerCase()) {
-            case "usd":
-                break;
-            case "eur":
-                expectedCost = expectedCost * 0.78460002;
-                break;
-            case "gbp":
-                expectedCost = expectedCost * 0.61250001;
-                break;
-            case "uah":
-                expectedCost = expectedCost * 28.22000000;
-                break;
-        }
-        String expectedPrice;
-        if (expectedCost > 1000) {
-            expectedPrice = String.format("%,3.0f", expectedCost);
-        } else {
-            expectedPrice = String.format("%.2f", expectedCost);
-        }
-        Assert.assertTrue(container.getPrice().contains(expectedPrice), "123");
-    }
-
     public HeaderPageBL clickOnLogoutButton() {
         headerPage.getLogoutButton().click();
         return this;
